Move header into EngFormat-Cpp directory
