+++ CMakeLists.txt
@@ -1,7 +1,7 @@
 project(EngFormat-Cpp CXX)
 cmake_minimum_required(VERSION 3.0)
 
-add_library(${PROJECT_NAME} include/eng_format.hpp src/eng_format.cpp)
+add_library(${PROJECT_NAME} include/EngFormat-Cpp/eng_format.hpp src/eng_format.cpp)
 
 target_include_directories(${PROJECT_NAME} PUBLIC
   $<BUILD_INTERFACE:${PROJECT_SOURCE_DIR}/include>
