+++ examples/example1.cpp
@@ -1,4 +1,4 @@
-#include "eng_format.hpp"
+#include <EngFormat-Cpp/eng_format.hpp>
 #include <iostream>
 
 int main()
