+++ src/eng_format.cpp
@@ -20,7 +20,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#include "eng_format.hpp"
+#include "EngFormat-Cpp/eng_format.hpp"
 
 #include <cctype>
 #include <cfloat>
